backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.11
httpx==0.25.2
python-dotenv==1.0.0
passlib==1.7.4
itsdangerous==2.1.2
python-multipart==0.0.21